fix:feature gate agntless and fix ubunut tests
