+++ datadog-sidecar-ffi/src/lib.rs
@@ -262,7 +262,6 @@ pub unsafe extern "C" fn ddog_remote_config_reader_for_endpoint<'a>(
             language: language.to_utf8_lossy().into(),
             tracer_version: tracer_version.to_utf8_lossy().into(),
             endpoint: endpoint.clone(),
-            agentless: None,
         },
         &Arc::new(Target {
             service: service_name.to_utf8_lossy().into(),
