+++ datadog-sidecar/src/service/sidecar_server.rs
@@ -779,7 +779,6 @@ impl SidecarInterface for ConnectionSidecarHandler {
                 language: config.language,
                 tracer_version: config.tracer_version,
                 endpoint: config.endpoint,
-                agentless: None,
             },
             products: config.remote_config_products,
             capabilities: config.remote_config_capabilities,
