+++ libdd-remote-config/Cargo.toml
@@ -11,6 +11,7 @@ description = "Datadog Remote Configuration client and config parsers"
 
 [features]
 default = ["client", "https"]
+agentless = ["tuf", "client", "https"]
 client = [
     "libdd-trace-protobuf",
     "http-body-util",
@@ -37,7 +38,7 @@ test = ["hyper/server", "hyper-util"]
 anyhow = { version = "1.0" }
 libdd-common = { path = "../libdd-common", version = "5.0.0", default-features = false }
 libdd-capabilities = { path = "../libdd-capabilities", version = "2.0.0" }
-libdd-capabilities-impl = { version = "2.0.0", path = "../libdd-capabilities-impl", features = ["https"]}
+libdd-capabilities-impl = { version = "2.0.0", path = "../libdd-capabilities-impl", default-features = false }
 libdd-trace-protobuf = { path = "../libdd-trace-protobuf", version = "3.0.2", optional = true }
 hyper = { workspace = true, optional = true, default-features = false }
 http-body-util = { version = "0.1", optional = true }
@@ -60,7 +61,7 @@ serde_json = { version = "1.0", features = ["raw_value"] }
 serde_with = "3"
 thiserror = "2"
 hashbrown = "0.15"
-tuf = { git = "https://github.com/DataDog/rust-tuf/", tag = "0.3.0-beta10-opw-3", default-features = false }
+tuf = { git = "https://github.com/DataDog/rust-tuf/", tag = "0.3.0-beta10-opw-3", default-features = false, optional = true }
 prost = "0.14.1"
 futures = { version = "0.3", features = ["executor"] }
 # `EnumIter` for external consumers to make struct available to runtime
