+++ libdd-trace-protobuf/build.rs
@@ -307,6 +307,10 @@ fn generate_protobuf() {
         "ClientGetConfigsResponse.client_configs",
         "#[serde(default)]",
     );
+    config.field_attribute(
+        "ClientGetConfigsResponse.config_status",
+        "#[serde(default)]",
+    );
 
     config.type_attribute("ClientUpdater", "#[derive(Deserialize, Serialize)]");
     config.type_attribute("PackageState", "#[derive(Deserialize, Serialize)]");
