+++ libdd-tracer-flare/Cargo.toml
@@ -35,4 +35,5 @@ tokio = { version = "1.36.0", features = ["time", "macros", "rt"] }
 
 [features]
 default = ["listener"]
+agentless = ["libdd-remote-config/agentless"]
 listener = ["libdd-remote-config/client"]
